(theme) Make Certain Styles Contained to Docs
The body of the docs subdomain has the class "chocolatey-docs". In some
cases, styles will not want to be used globally, so to narrow the scope
to a certain repository, this class can be used. Each repository will
have its own specific class.

In this case, a few styles that are hiding/showing specific elements
have been narrowed down to the docs subdomain.