(theme) Add Support for Carousels
This adds support for carousels that can be used at a global level.
Specific JS has been added to initialize the carousels, and prevent
it from looping on autoplay on mobile. Sometimes this can cause the
page to jump when the container is not at a fixed height, which can be
very user unfriendly on mobile.