(theme) Carousel Additions and Improvements
An added type of carousel has been introduced, one that contains only
images in a block style. Since this type of carousel will always be the
same height (assuming the images are the same size), the JS has been
reworked to not include this and start automatically on load (based on
markup).