(GH-17)(GH-18)(GH-19) Bootstrap v5 & JS Updates
Upgrades to Bootstrap v5. This included lots of changes
including lost support for IE, and dropping jQuery as a dependency.
As a result, many of the custom JS functions have been reworked to
use basic Javascript as opposed to jQuery. To enable support for IE,
several styles have been added to an IE specific scss file, as well
as JS polyfills.

The JS files have been split into many different files to allow the
theme to be more specific on what is included, and reduce file size.
This also helps with debugging and reduces conflicts.