(js) Add Observer for Document Height Change
In some cases the window resize event is not sufficient for determining
the left side bar height, resulting in a sidebar that does not connect
to the bottom of the page. The observer watches for changes in the
document height, and at that point will be triggered. This will ensure
the sidebar always has the correct height.