(js) Add URL Hash/Collapse/Tab Toggle
This adds back in previous functionality that changes the url hash when
a collapsed/tabbed element is clicked on. When navigating to a url that
contains a hash, the JS will look to see if the hash is a
collapsed/tabbed item, and if so it will open the collapse/tabbed
element to be viewed.