(events) Update Countdown for Reoccuring Events
Adds support into the countdown timer on the event pages for weekly
reoccuring events. Now, whenever the event is over, the timer will
automatically lapse over to the next weekly date. The new date will
also be displayed on the page so it is easier to tell when the next
webinar is rather than just "Every Thursday".

Support will need to be added in if there is a different interval
needed other than weekly.