(build) Pin easymde to v2.13.0
The easymde package needs to be pinned to version 2.13.0. This is
due to the fact that newer versions have a dependency on the package
"types/marked v2.0.3" which depends on a higher version of node than
what is installed on Appharbor. Pinning easymde to version 2.13.0 only
depends on "types/marked v1.2.3" which is supported by the currently
installed version of node.