(packages) Add Color for Scan Exemption Note
The newest release of Chocolatey has the scanner status of exempted with
a note. This type of status has not been accounted for in the css, and
the icon next to the package did not have a color. The appropriate gray
color has been added for the sliver of the icon, which matches the
exempted icon in on the package page.