(packages) Add Color for Investigate Note
The newest release of Chocolatey has the scanner status of investigate
with a note. This type of status has not been accounted for in the css,
and the icon next to the package did not have a color. The appropriate
red color has been added for the sliver of the icon, which matches the
icon in on the package page.