(script builder) Remove "--pre" from Title
Currently, if a package is added to Script Builder that has the --pre
flag, it is being added to the title of the package in the generated
script. This ensures that the value is pulled out from the start of the
first `--` character and split, regardless of what flag is listed.