(#131) Update CSS/JS for sticky-top
The CSS for determining the position of the sticky-top navigation
needed to be a bit smarter. If the sticky top navigation is not enabled,
then the any div with the "sticky-top" class should be positioned at
0. If the sticky top navigation is enabled however, any other div with
the class "sticky-top" should be set at 65px (height of top nav). This
allows for correct positioning of both the top nav and any additional
sticky items on the page.