(tables) Add styles to support sticky table head
Creating a responsive table with a sticky header is tricky. A few
custom styles have been added to support this. A simple JS script has
also been added which will scroll the sticky header along with the body
of the table horizontally. Certain utility classes are also defined
here that can be used on tables, or wherever needed.