(#138) Script Builder- Compare Package Id vs Title
Previously, Script Builder was comparing the package Title when trying
to determine if there was already a package with the same name in a
users builder. In most cases this works, but in the case that the
package title changed, it was allowing multiple versions of the same
package and not prompting a warning to choose one.

Now, instead of comparing package Title, the package Value (Id) is
compared. This is a more reliable way of knowing if there is already a
version of the same package in a users builder since the Id of the
package will never change.