(#142) Navigate to nested items on hash change
The implementation for navigating to a hash that located inside of a
collapsed or tabbed element already existed, however it was commented
out. This removes the comments to enable this functionality.

This will mainly be used on the docs site when clicking on a right hand
navigation item that may be located inside a collapsed element.