(maint) Small Updates to Capitalization
This updates the capitalization of some classes to help them be
compatible out of the box with classes that are added with JS without
needing to convert to lowercase.

A new class has also been added to manually add an underline to a div
like it would be added to an h1 element. This is a very edge case
scenario.