(#151) Ensure Luxon Date Conversions are Correct
To ensure that dates converted by Luxon are correct, changes to the
initial input string were made along with some refactoring of the JS
that controls the conversions.

The default string that is captured from the html has been updated from:
07/29/2020 03:30:00

To:
2020-07-29T03:30:00Z

Updating the string ensures that conversions will be done correctly
between time zones and will accurately take into account and offsets due
to daylight savings time.