Introduce Add to Calendar Button
A new JS library has been introduced that allows the creation of an
"Add to Calendar" button. This supplies many different formats and
calendar types that will likely fit most needs. This will enable a user
to add the event to their calendar. The "location" in the event invite
will always be the register url, just in case the user had not
registered yet.

The scss and js were modified slightly from their original state to fit
the needs of our use case. Some of the icons were replaced with font
awesome and the styling has been adjusted to fit into choco-theme. Lots
of the original bits have been left untouched, so hopefully upgrading
won't be too hard if that is needed in the future.