(#155) Create Partials for Terms of Use
This creates 3 different partials for the Terms of Use content. This
was split up because the content is needed on different parts of the
page, and some may not be needed at all.

Since the same Terms of Use are on all the websites, putting them in a
global partial is the best option. This way the content all stays in
sync when it is updated.