Prevent Button Text Change in Flyout
When an event is over, the button text is changed by JS automatically.
In the sidebar, this is not ideal, or hasn't been considered yet in the
JS, and was happening inaccurately. If on a an event page that was
expired, every button in the right flyout would say "Watch On-Demand"
even if the event hadn't happened yet. This update changes the buttons
in the flyout to be more specific, that way they can be excluded in the
JS.