Add lite-youtube-embed JS
This library helps with lazy loading of YouTube videos so that pages
with lots of videos on them don't get loaded until they are click on to
be played. This is slightly better than just lazy loading the video on
scroll, as nothing but a placeholder image will display that looks like
a video until it is actually really do be watched.