(#203) Adjust Zendesk Toggle Switch
This adjust the css used that will control to theme toggle switch on
the Zendesk Help Guide. Because of how assets are used and uploaded to
Zendesk, a unique approach was needed in order for the images to show
correctly. This will ensure the toggle switch functions and looks how
it should.