(#203) Updates to support Zendesk choco-theme
A new JS file has been added that will add the necessary Bootstrap classes to select inputs. This file may be amended to in the future to also select any input type.

As a result of adding a new file and also choco-theme to Zendesk, the README.md file has been updated.