(#228)(#210) Modernize JavaScript and ESLint
The JavaScript in choco-theme has been overhauled to use specific ES6
and higher syntax and limits the usage of jQuery.

The conversion of the JavaScript was done along side the introduction
of ESLint. This was done to help with enforcing specific formats for
writing functions to keep a consistent style across the board.

There is a new .eslintrc.js file that contains all the rules introduced
so far. When writing new JavaScirpt, warnings and errors will be thrown
if something is not in the correct format. This is not currently
built in to any pipelines or pre-commit hooks, although it will be in
the future.

ESLint can be ran in choco-theme by the following command:

`yarn run eslint yourfile.js`