(#235) Package Differ CSS and JS
This adds the initial CSS and JS needed by Package Differ. The prism.js
library has been updated to include the diff plugin and also a better
approach has been taken with the CSS. Now, the base CSS is imported by
prism.js and then overridden with custom styles. In most cases it's not
great to override styles, but here we want to do this in order to make
updates easier. A separate prism-overrides file is now present that
contains style and color choices specific to Chocolatey.