(#212) Upgrade to Bootstrap 5.2.2
With the recent upgrade to JavaScript, this allowed the Bootstrap
upgrade fairly straightforward. A change was made to set the main
border radius' back to how they have been determined in previous
versions of Bootstrap, since this change is not desired. Other than
that, nothing should really change look/feel wise on the front end.