(#259) Add reusable function to truncate results
In the packages area, there is a table with several results that should
be truncated if the results are over a certain number. A new function
has been added that can be referenced for instances like this by adding
the class 'truncate-results` to the rows that are intended to be
truncated.

The ID of the button should be passed into the function for the click
handler to be attached.