(#261) Ensure proper indention of Scripts
In Script Builder, there are a number of generated scripts that are
based on user input. These scripts were not indented properly. The
changes here ensure proper indentation for installation usage. A new
global function has also been added that will remove blank line breaks
from a string. This is useful in situations where template literals are
used on separate lines that may or may not contain content.