(#224) Make callouts more visible
Often times, callout (alert) are ignored. This becomes a problem when
the information in the callout is crucial. The changes in this commit
update the styling to the callouts to be more vibrant and eye catching,
so that hopefully they won't be missed so easily.

This also moves away from using emojis and uses Font Awesome icons
instead. This was done because in some cases, the emojis were hard to
see and not very accessible. Switching to Font Awesome allows them
to be styled in a way that is easier to see, and also brings consistency
to the rest of the style in choco-theme. Information on how to use these
new styles is in the design system.