(#270) Update design/CSS/JS based on choco-design
During the writing of the choco-design system, it because clear that
some styles in choco-theme were not very consistent across the board.
Changes have been done to bring that consistency to things like border
radious on components, shadows vs. no shadows, and also the naming of
classes.

Many class names and variable names have been changed to stay inline
with each other. This will bring clarity to new users who are hoping to
become familiar with choco-theme.