(#239) Introduce Stylint
Stylelint has been added into choco-theme with a set of rules that is
very similar to what Bootstrap uses. If a rule is broken, an error
will occur in the users IDE (only test on VSCode at this time). A
stylelint config has been added and it can be ran from the command line.
The README.md file has been updated to include further instructions on
how to run Stylelint to find and fix potential errors.