(#245) Offcanvas for Flyouts
The left and right side flyouts have been replaced with the Offcanvas
component, rather than the Collapse component. The Offcanvas component
offers smooth transitions and backdrops that improve the usability of
these functions. Overall the functionality is the same, but enhanced.
The CSS and JS were updated slightly to implement this feature
correctly.