(#293)(#294) Revamp Login and Account Navigation
The login and account buttons on community.chocolatey.org have been
moved into the main navigation area. Having them in this area makes the
experience more intuitive and familiar. Navigating to the Account page
from the first page of packages (or any page) will now take the user to
the correct page instead of a broken link. The url for the account page
is now being inserted with JS when the user is authenticated instead of
in the C# page. Due to caching, this needs to be done on the client
side to ensure the proper link is available.