(#296) Reoccurring Add to Calendar Buttons
This updates the library used for the Add To Calendar buttons
and enables the usage of reoccurring events. The package
for this library has been installed, and a override scss file has
been placed. This will ensure upgrading of the library is much
easier in the future. In each repository that this is implemented on,
the css is imported with the bundleconfig.json file, and the JS is
initialized in the init.js.