(#297) Update Product Spotlight Event
This adds a global partial for the Chocolatey Product Spotlight
event. The partial contains an Add to Calendar
button that will be used in various places on chocolatey.org
and community.chocolatey.org. When the dates of this event
need to be updated, they should be changed here so that the
calendar invite dates are changed.

This partial is imported into the right side bar via the gulpfile.js
in chocolatey.org and community.chocolatey.org.