(#296)(#297) Allow Buttons to Wrap in Flyout
With the addition of the Add to Calendar buttons, there are now
two buttons in the flyout next to each other. A new class has been
added to the containing div to allow to buttons to wrap properly
on smaller screens.

In addition, a small enhancement to the JS has been made to not
remove the Add to Calendar buttons on an event page after the event
time is over. This is due to the Add to Calendar buttons now containing
more than just one event date.