(#301) Provide auto close modal functionality
By adding the class "modal-auto-close" on the modal, and then
"modal-auto-close-btn" on the trigger, the modal will close. Most of
the time the "modal-auto-close-btn" would be intended to be used on an
external link. This will ensure that when the user comes back to the
Chocolatey page, that the modal is still not open. This will reset the
user back to where they expect to be.