(#311) Update Home Page Design
This adds new and adjusts css and js to give the home pages a fresh
new look. This adds a new carousel library that is used for autoscroll
behavior, since this is not a feature with Bootstrap's carousel.

Another notable thing is the change to the border radius on cards. This
has been increased to 1rem instead of .25rem, which is a noticeable
change. This will also take effect on all cards across all the sites,
not just the home pages. This border radius change is likely to come
to images and callouts in the future for consistency.