(#309) Ensure Markdown Callouts Always Render
In the JS, we are not checking for the existence of the default `<p>`
element before we check for it's text contents. This was throwing an
error if the callout did not contain a `<p>` element to begin with. This
mostly will happen if the callout contains html elements on the first
line.