(#311) Minor adjustment to home page updates
These scss here has been modified slightly to account for a few minor
adjustments that are needed after deploying the updates to the home
page.

A -webkit property has been added so that the background-filter
property works in Safari. In addition, the font sizes of the
typing terminal commands has been adjusted slightly to ensure it doesn't
wrap around and make the screen jump.