(#316) Update to DocSearch to V3
This upgrades DocSearch (Algolia) to V3. The search is no controlled by
a dedicated search box snippet provided by Algolia and, for the most
part, overridden by css variables supplied by them.

The search is now inside of a modal as there is no option to have it in
a dropdown like before.