(#332) Remove extra space in generated xml file
After a bit of debugging, it was found that the problem of the extra
leading space was not in the generation of the file, but in how it was
being downloaded. When setting the `href`of the created element, there
was an extra space before the text was included. Normally, extra spaces
like this are not a problem, but because of the type of function being
performed, this extra space made it's way into the downloaded xml file.

By removing this space, there is now no leading white space in the
generated xml file from Script Builder.