(#285) Pin Bootstrap to 5.2.3
A new stable version of Bootstrap, 5.3.0, has been released, however this does
cause a few things to break in choco-theme. It is best to pin Bootstrap to the
last known version to work, 5.2.3, until a proper upgrade can be done. If this
version is not pinned, then 5.3 will be pulled automatically on upgrade of
choco-theme, which is not intended.