(#336) Adjust cards to not show overflowing content
In a previous version of choco-theme, the border radius of
cards was increased. Unfortunetely, this brought on some
unwanted side effects, including some content that appeared
to spill over the card. Instead of manually going through every
card and added a "overflow-hidden" class to it, it has been
changed in the main CSS. There are a few one off instances
where the overflow needs to be visible, and in those cases,
the "overflow-visible" class has been added to them.