(#340) Update ESLint GitHub Action on pull request
Since it was proven that stylelint works well with a more
direct approach, the ESLint action has been updated to
simply run the appropriate terminal command to check for
errors in all js and ts files. If there are any, the check
will fail, and the errors are stated in the Action log.

In order for this to work, TypeScript was downgraded to the
last stable version instead of the most stable. The ESLint
parser was not compatible with the current version of
TypeScript. This should not pose any problems, as it will
likely catch up soon and TypeScript can be updated again.