(#335) Add Multiple Tab Support
Sometimes, multiple tabs in different spots of the page need triggered
at the same time. An example of this use case would be on
docs.chocolatey.org and separating Chocolatey v1 information from
Chocolatey v2.

This is achieved by adding the data-choco-tab-multi='{"${cookieName}":
"${cookieValue}"}' attribute to each .nav-link with the same
corresponding value of those that need to be triggered.