(#347) Add AnchorJS anchors to list items
This adds anchors to list items that contain a <strong> or a <code>
node directly after an opening <li> tag. This is to make it easier to
navigate to these items, as often times this format is used more as a
"definition", or something we'd want to potentially link to.

The order of the JS imports needed to be modified to ensure that the
markdown conversion JS for list items was complete before trying to
detect the desired nodes inside the list items.