(#360) Styles to Support the Use Cases > Server
The changes here are used and will be used on
the new Use Cases > Server page, and future
additional pages.

Introduced is a CSS pollyfill that will balance
text automatically based on the screen width.
In the future, once this is supported in all
browsers with CSS, this can be removed.

A few new background colors have been added for
usage on icons to better outline problem/solution
paragraphs.