(#392) Remove Gulp
This removes Gulp and introduces a number of new
technologies. The build process is now handled by
esbuild and a number of new scripts. The build
process is now much faster and more reliable with
less chance of introducing security
vulnerabilities.

Choco-design has been updated to reflect these
changes as well as the README of this repository.