(#392)(#285)(#218) Upgrade atcb.js
As a result of upgrading Yarn, this library was
reported to be deprecated and noted that it may
cause security vulnerabilities. This PR upgrades
the library to the latest version. Many changes
were needed as this was a major version change.
Now, the button pulls in a custom css file
directly from the cshmtl instead of it being
bundled with the rest of the css. This is
following best practices for the library.