(#392)(#285) Updates to README and package.json
These are updates to the scripts in the package.json
so that they can be ran with both yarn and npm, with the
preferred method being yarn. This also moves some of the
packages into devDependencies, so that they are not installed
on each individual repository that uses choco-theme. They
are only needed when building the initial bundles.