(#392) Update safelist for community
This updates the safelist for the community repository when PurgeCSS is
ran. Before, we were able to look at the node_modules folder to compare
content, but now that most libraries are dev dependencies of
choco-theme, they are not installed on each repository. In this case,
they are no longer there to be looked at. The fix for this is to,
instead look at the source CSS, to find the top level CSS locater and
add it to the safelist of things not to be purged. By doing this, all
the necessary CSS shows up in the final purged CSS file.