(#400) Allow GitHub Action to publish to NPM
As found in the Yarn docs, by adding a
`-fallback` onto the passed in environment
variable, if one is not supplied, it will
"fallback" to a default value. This is exactly
what is needed in order for repositories to use
choco-theme that do not have this environment
variable set. This will allow GitHub Actions to
publish to NPM and also allow repositories to
install choco-theme like normal.