(#405) Support Astro Framework
The changes here are needed in order to support
the Astro framework. A few styles were updated,
but the majority of the changes are in the build
process and some updates to converted markdown
inside code blocks. All existing choco-theme
commands still work and are ran the same way from
the command line, except now when running anything
with `--docs` it will run Astro commands instead
of a preview.sh or preview.ps1 file.