(#412) Copy bootstrap.bundle.min.js
Instead of importing Bootstrap into each file,
Bootstap will now be its standalone file that is
copied over to each repository from node_modules.

Due to Bootstrap needing to be used in the global
scope now, this was causing issues with the
eslint and esbuild portions of the build process.