(#412) Update build pipeline for CCM
This updates the build pipeline slightly for the CCM
project. Instead of the dist folder being cleaned in
choco-theme, it is now cleaned in a file in the repository
itself due to race conditions and ordering of scripts being
executed. A line was added to the purgecss config to ensure
that any styles related to syntax highlighting are not purged.
In addition, an extra console.log has been removed that is
unneeded.