(#417) Update yarn npm audit command
This updates the yarn npm audit command to include all severity
types, scan dependencies recursively, and also ignores the known
vulnerability in sweetalert2.

This command has been added to the readme, which can be ran locally
after `yarn` has been ran. In addition, this has been updated in the
audit.yaml file that is ran before choco-theme is released.