(#418) Upgrade to ESLint 9.x
This upgrades to the latest version of ESLint. With this upgrade, the
format of the config file change to a `eslint.config.mjs` file. A tool
was used by ESLint to convert over the existing .eslintrc.js file. More
information on the conversion can be found at https://eslint.org/docs/latest/use/configure/migration-guide.

After this upgrade of ESLint, ESLint was ran and tested by making
purposeful changes to ensure that errors were caught. This worked
successfully, and also found an actual error that was fixed.