(#422) yarn preview-upgrade improvements
This adjust the yarn preview-upgrade command to first remove the
yarn.lock file in the chosen repository before upgrading choco-theme.
By doing this, it ensures all dependencies are upgraded of choco-theme
and not just choco-theme itself. This is desired in most cases.

Also, the CCM repository has been added to the list of configs, so that
is caught when using the --all or --ccm flag on yarn commands.