(#427) Add Tab styles to copenhagen
This adds styles for the Tab component to the copenhagen theme css.
On most repositories, classes are used on the component to define these
styles, but, to ease development on this specific repository the styles
have been added directly to the css. This is so that examples can be
copied directly from the Bootstrap docs, and still be styled up to
choco-theme.