(#453) Enhance left side bar menu in Chocolatey Central Mangement
This adds some extra padding on nested list items in the left menu, so
that they appear correctly nested under the parent. This also increases
the max width of the sidebar to account for the extra padding, and allow
for items not to wrap on large screens. This also allows all links to be
visible when tabbing for all browsers.