(#454) Allow dark mode colors to be used in light mode
The changes here allow for a specific element to be in "dark mode" while
the rest of the page is in "light mode". This can be achieved by adding
the correct `data-bs-theme` onto an element itself. This is a supported
feature by Bootstrap, but since we have added to the root element, we
also now need to support it in choco-theme.