(#444)(#456) Enhance eslint and apply rules
This updates eslint to the latest recommended
config via the eslint documentation. This also
ensures that this eslint config can be used on
most current projects that utilize eslint. This
adjusts the choco-theme command to include the
choco-playwright repository as a website, so that
the config may be copied and utilized there. This
also applies the eslint rules and adjusts some
files to make sure all the rules are followed.