(#460) Craft a credits.json file shell
The changes here are the begining steps to crafting a credits.json
file. This file will be used to later generate an updated version
of the CREDITS.md. For now, the credits.json file can be created by
running `yarn credits`. Right now, links to the license files are
missing for each dependency. This is a work in progress that will be
added to in a later commit.