(#460) Html to pdf and md
This adds a few new packages to allow the
conversion of HTML to PDF and Markdown. While right
now this is only being used on one website, it has
been added to choco-theme so other repositories
can use it in the future. On the website that
uses this, there are export buttons that when
clicked, run the helpers here to generate the
files.