(#440) Dynamic code block functionality
This allows for the dynamic code block component in choco-astro to
listen for values entered into input for code block components and
replace custom variables placed in the code block on a page.

This also fixes a small bug, where the "copy" button on dynamic code
blocks was not working properly and either not copying text, or not
updating the button text.