(#440) Fix dynamic code block issue with regex
Added regex to find exact match when doing a replace all in the dynamic code block. This fixes issue when multiple inputs are used on a page and one input name or variable has no number but other inputs do have a number in the name attribute variable it doesn't break the find and replace functionality.