(#440) Add cookies for Dynamic Code Blocks
This adds the ability to set and read cookies to the Dynamic Code Blocks
which will help so that users don't have to enter the same information
on multiple screens.

In addition the set and get cookie functions were updated slightly to
allow the usage of all special charachters by first escaping them, this
was not previously supported.