(#460) Utils to support credits generation
These utils will help support the generation of credit files, however
they can be used in multiple projects for other reasons. The to-id
function converts a string to a valid Id. The get-unique-id function
ensures there are no duplicate Id's on a given page, and follows the
common format of adding numbers, like in a toc, to the end. The
regenerate-modal function will trigger a "loading" or "regenerating"
modal on a button click. This is useful to stop user interaction
while a page is "thinking".