(#460) Address edge cases in credits.json
There are a few packages that are "edge" cases, in that maybe the
versions are odd, or the links need manually inserted. This commit
addresses a few of these so that the Credits Generator can properly
find all license links and the .json file is valid.

This updates the capitalization of the credits.json file to CREDITS.json.
This change was done due to the generated .pdf and .md files being
capitalized as well, for consistency. This will help to easily group
and identify these files.