(#468) Bump choco-theme to 1.3.1
This bumps choco-theme to 1.3.1. Included in this release are the
upgrades of dependencies relating around eslint and the typescript
packages. For now, typescript has been pinned to the version it is on
because there are compatibility issues with the latest version with
eslint.

This also upgrades dependencies for Stylelint. With the upgrade of these
dependencies, it has automatically removed any depreciation errors that
were showing up when running stylelint.