(#467) Create @chocolatey-software/build-tools
A new package has been created @chocolatey-software/build-tools to
contain shared build tools for other packages. The build scripts in
this folder, along with the dependencies in package.json, are ones that
each repository needs access to at build time. This package will be
installed as a devDependency in other packages to provide those tools.

A handful of scripts have also been removed to clean up unused
functionality that will no longer work with the current architecture
without additional work. These can be re-added in the future if
needed.