(#467) Update package.json
Since a lot of the changes to the package.json were done in bulk when
splitting up packages to workspaces, this commit just updates the
this main package.json to reflect the new versions of the packages.
This also brings in the new commands that can be ran in order to build
and lint all the various packages.