(#477) Hide or disable Tabs component
This adds two more options that can be passed into the tabs component
in @chocolatey-software/astro: isDisabled, and isHidden.

If the isDisabled option is passed in on a tab with "true", then the tab
heading will show, but the user is unable to click or navigate tot he
content. The content is however still available in the dom, if a user was
to inspect the source code.

If the isHidden option is passed in on a tab with "true", then the tab
and the content pane will not be rendered on the page at all.

Thse two options offer various ways of showing "work in progress" without
showing the actual content just yet.