(maint) Suppress if() SASS depreciation warning
There is a new SASS deprecation warning for using the if() function.
This is coming from files within Bootstrap itself, and we should wait
for Bootstrap to fix this in a later version before we try to
override/or not use it. Right now, we are suppressing the warning to not
draw attention away from actual errors.