(#490) Add identifier to Playwright axe flow
This adds the identifier `#SendActivationEmailHelp` to the list of
identifiers to ignore on axe color checks. This has been added due
to it reporting a false positive of color contrast. It states the
identifiers background color can not be determined due to overlapping
elements. We know for a fact here that this element does not have a
set background color, and is text on the modal background. This element
passes manual color checks, and therefor is ignored here.